+++ futurehome/src/ha/mqtt_components/_enums.ts
@@ -1,1 +1,14 @@
 export type EntityCategory = undefined | 'config' | 'diagnostic';
+
+export type CoverDeviceClass =
+  | 'awning'
+  | 'blind'
+  | 'curtain'
+  | 'damper'
+  | 'door'
+  | 'garage'
+  | 'gate'
+  | 'shade'
+  | 'shutter'
+  | 'window'
+  | null;