+++ futurehome/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added support for 'indicator_ctrl' service (identify devices).
 - Added support for 'barrier_ctrl' service (devices like garage doors, barriers, and window shades).
 - Added support for 'water_heater' service (devices such as water boiler or a water tank).
+- Updated demo mode data.
 
 # 0.1.0 (24.07.2025)