+++ futurehome/CHANGELOG.md
@@ -1,6 +1,12 @@
 <!-- https://developers.home-assistant.io/docs/add-ons/presentation#keeping-a-changelog -->
 
 
+## 0.1.2 (24.07.2025)
+
+- Added support for 'water_heater' service (devices such as water boiler or a water tank).
+- Updated demo mode data.
+- Added extracting device manufacturer name.
+
 ## 0.1.1 (24.07.2025)
 
 - Set 'battery' entity category to 'diagnostic' to hide it from the main HA view.
@@ -8,9 +14,6 @@
 - Changed the default 'sensor_lumin' unit from 'Lux' to 'lx'.
 - Added support for 'indicator_ctrl' service (identify devices).
 - Added support for 'barrier_ctrl' service (devices like garage doors, barriers, and window shades).
-- Added support for 'water_heater' service (devices such as water boiler or a water tank).
-- Updated demo mode data.
-- Added extracting device manufacturer name.
 
 # 0.1.0 (24.07.2025)